suggestion from code review
